school/requirements.txt
asgiref==3.2.7
Django==3.0.6
django-cors-headers==3.3.0
djangorestframework==3.11.0
kids==1.0.4
Pillow==7.1.2
pytz==2020.1
redis==3.5.2
sqlparse==0.3.1
